separate bot from app
